+++ lib/http2/browser/browser.js
@@ -1,14 +1,14 @@
-import { Http2WebTransportSession } from './session.js'
+import { Http2WebTransportSession } from '../session.js'
 import { BrowserParser } from './browserparser.js'
 
 export class Http2WebTransportBrowser {
   /**
-   * @param {import('../types').NativeClientOptions} args
+   * @param {import('../../types.js').NativeClientOptions} args
    */
   constructor(args) {
     this.port = args?.port || 443
     this.hostname = args?.host || 'localhost'
-    /** @type {import('../session').HttpClient} */
+    /** @type {import('../../session.js').HttpClient} */
     // @ts-ignore
     this.jsobj = undefined // the transport will set this
     /** @type {WebSocket} */
