+++ lib/http2/browser/browserparser.js
@@ -1,4 +1,4 @@
-import { ParserBase, lengthVarInt } from './parserbase.js'
+import { ParserBase, lengthVarInt } from '../parserbase.js'
 
 /**
  * @param{{offset: Number, buffer: Uint8Array, size: Number}} bs
@@ -54,7 +54,7 @@ export class BrowserParser extends ParserBase {
   static WS_PING = 0x9
   static WS_PONG = 0xa
   /**
-   * @param {import('../types').ParserWebsocketInit} stream
+   * @param {import('../../types.js').ParserWebsocketInit} stream
    */
   constructor({ ws, nativesession, isclient }) {
     super({ nativesession, isclient })
