+++ lib/http2/node/capsuleparser.js
@@ -1,9 +1,9 @@
-import { ParserBaseHttp2, readVarInt, writeVarInt } from './parserbasehttp2.js'
-import { lengthVarInt } from './parserbase.js'
+import { ParserBaseHttp2, readVarInt, writeVarInt } from '../parserbasehttp2.js'
+import { lengthVarInt } from '../parserbase.js'
 
 export class Http2CapsuleParser extends ParserBaseHttp2 {
   /**
-   * @param {import('../types').ParserHttp2Init} stream
+   * @param {import('../../types.js').ParserHttp2Init} stream
    */
   constructor({ stream, nativesession, isclient }) {
     super({ stream, nativesession, isclient })
