+++ lib/http2/node/client.js
@@ -1,7 +1,7 @@
 import { connect } from 'node:http2'
 import { Http2WebTransportSession } from './session.js'
 import { log } from 'node:console'
-import { Http2CapsuleParser } from './capsuleparser.js'
+import { Http2CapsuleParser } from './capsule/capsuleparser.js/index.js'
 
 export class Http2WebTransportClient {
   /**
