+++ lib/http2/node/server.js
@@ -1,5 +1,5 @@
 import { createSecureServer, constants } from 'node:http2'
-import { Http2WebTransportSession } from './session.js'
+import { Http2WebTransportSession } from '../session.js'
 import { Http2CapsuleParser } from './capsuleparser.js'
 import { WebSocketParser } from './websocketparser.js'
 import { log } from 'node:console'
@@ -7,7 +7,7 @@ import { webcrypto as crypto } from 'crypto'
 
 export class Http2WebTransportServer {
   /**
-   * @param {import('../types').NativeServerOptions} args
+   * @param {import('../../types.js').NativeServerOptions} args
    */
   constructor(args) {
     let port = args?.port
@@ -27,7 +27,7 @@ export class Http2WebTransportServer {
     /** @type {Record<string, boolean>} */
     this.paths = {}
     this.hasrequesthandler = false
-    /** @type {import('../session').HttpServer} */
+    /** @type {import('../../session.js').HttpServer} */
     // @ts-ignore
     this.jsobj = undefined // the transport will set this
 
@@ -277,7 +277,7 @@ export class Http2WebTransportServer {
   }
 
   /**
-   * @param {import('../types').NativeFinishSessionRequest} args
+   * @param {import('../../types.js').NativeFinishSessionRequest} args
    */
   finishSessionRequest({ header, session: stream, status, protocol, head }) {
     if (status !== 200) {
