+++ lib/http2/node/websocketparser.js
@@ -1,6 +1,6 @@
 import { randomBytes } from 'node:crypto'
-import { ParserBase, lengthVarInt } from './parserbase.js'
-import { ParserBaseHttp2, readVarInt, writeVarInt } from './parserbasehttp2.js'
+import { ParserBase, lengthVarInt } from '../parserbase.js'
+import { ParserBaseHttp2, readVarInt, writeVarInt } from '../parserbasehttp2.js'
 
 /**
  * @typedef {import('node:http2').Http2Stream} Http2Stream
@@ -140,7 +140,7 @@ export class WebSocketParser extends ParserBaseHttp2 {
   static WS_PING = 0x9
   static WS_PONG = 0xa
   /**
-   * @param {import('../types').ParserHttp2Init} stream
+   * @param {import('../../types.js').ParserHttp2Init} stream
    */
   constructor({ stream, nativesession, isclient }) {
     super({ stream, nativesession, isclient })
