+++ lib/server.js
@@ -1,7 +1,7 @@
 import { ReadableStream } from 'node:stream/web'
 import { HttpWTSession } from './session.js'
 import { WebTransportError } from './error.js'
-import { Http2WebTransportServer } from './http2/index.js'
+import { Http2WebTransportServer } from './http2/node/index.js'
 import { isIPv4 } from 'net'
 import { defer, logger } from './utils.js'
