+++ lib/webtransport.browser.js
@@ -1,7 +1,7 @@
 import { WebTransportBase } from './webtransportbase.js'
 import { HttpWTSession } from './session.js'
 import { HttpClient } from './client.js'
-import { Http2WebTransportBrowser } from './http2/browser.js'
+import { Http2WebTransportBrowser } from './http2/browser/browser.js'
 
 /**
  * @typedef {import('./dom').WebTransport} WebTransport
