+++ lib/webtransport.node.js
@@ -1,7 +1,7 @@
 import { HttpWTSession } from './session.js'
 import { WebTransportBase } from './webtransportbase.js'
 import { HttpClient } from './client.js'
-import { Http2WebTransportClient } from './http2/index.js'
+import { Http2WebTransportClient } from './http2/node/index.js'
 import { logger } from './utils.js'
 
 const log = logger(`webtransportnode(${process?.pid})`)
